requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
cloudinary==1.36.0
openpyxl==3.1.2
reportlab==4.0.8
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1